Add files via upload
The project presents a classification machine learning model to classify restaurant reviews. The reviews can be anything which are related to the food of the restaurant, staff of the restaurant and also overall review of the restaurant. The model uses Support Vector Machine (SVM) algorithm for classifying the reviews. The classified reviews are helpful for the restaurant to analyze their shortcomings in different areas and improve the quality of food and service in the restaurant.